requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.6
django-environ==0.9.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dnspython==2.3.0
environ==1.0
future==0.18.3
idna==3.4
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
oauthlib==3.2.2
openapi-codec==1.3.2
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
simplejson==3.18.3
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
